Custom EventSource working

+++ DiagExperimentsSolution/TestConsole/Helpers/DumpHelper.cs
@@ -79,7 +79,7 @@ namespace TestConsole.Helpers
             var token = _tokenSource.Token;
             _gcroot = new GCRoot(_clrRuntime.Heap);
 
-            _gcroot.ProgressUpdated += delegate (GCRoot source, int processed) //long current, long total)
+            _gcroot.ProgressUpdated += delegate (GCRoot source, long processed) //long current, long total)
             {
                 //var percent = total == 0 ? 0 : (int)(100 * current / (float)total);
                 //OnGCRoot?.Invoke((current, total, percent, token));
